+++ src/main/java/com/qlyshopphone_backend/dto/NotificationDTO.java
@@ -7,9 +7,9 @@ import java.time.LocalDateTime;
 
 @Data
 public class NotificationDTO {
-    private Integer notificationId;
+    private Long notificationId;
     private String message;
-    private Integer userId;
+    private Long userId;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime timestamp;
 }
