+++ src/main/java/com/qlyshopphone_backend/dto/PurchaseDTO.java
@@ -7,10 +7,10 @@ import java.time.LocalDateTime;
 
 @Data
 public class PurchaseDTO {
-    private Integer purchaseId;
-    private Integer userId;
-    private Integer customerInfoId;
+    private Long purchaseId;
+    private Long userId;
+    private Long customerInfoId;
     private BigDecimal totalPrice;
-    private Integer totalAmount;
+    private Long totalAmount;
     private LocalDateTime purchaseDate;
 }
