+++ src/main/java/com/qlyshopphone_backend/model/Properties.java
@@ -11,7 +11,7 @@ public class Properties {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "properties_id")
-    private Integer propertiesId;
+    private Long propertiesId;
 
     @Column(name = "properties_name", length = 50, nullable = false)
     private String propertiesName;
