+++ src/main/java/com/qlyshopphone_backend/model/Trademark.java
@@ -11,7 +11,7 @@ public class Trademark {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "trademark_id")
-    private Integer trademarkId;
+    private Long trademarkId;
 
     @Column(name = "trademark_name", length = 50, nullable = false)
     private String trademarkName;
