+++ src/main/java/com/qlyshopphone_backend/model/UserStatistics.java
@@ -13,7 +13,7 @@ public class UserStatistics {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_statistics_id")
-    private Integer userStatisticsId;
+    private Long userStatisticsId;
 
     @OneToOne
     @JoinColumn(name = "user_id", nullable = false)
@@ -23,5 +23,5 @@ public class UserStatistics {
     private BigDecimal totalAmountPaid;
 
     @Column(name = "total_item_bought", nullable = false)
-    private Integer totalItemBought;
+    private Long totalItemBought;
 }
